requirements.txt
streamlit
pytesseract
Pillow
pdf2image
pay-by-square
qrcode
numpy
